Add product cart

+++ Shop2/Data/Models/Car.cs
@@ -12,7 +12,7 @@ namespace Shop2.Data.Models
         public string ShortDesc { get; set; }
         public string LongDesc { get; set; }
         public string Img { get; set; }
-        public ushort Price { get; set; }
+        public uint Price { get; set; }
         public bool IsFavourite { get; set; }
         public bool Available { get; set; }
         public int CategoryID { get; set; }
